% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_RoBSA}
\alias{check_RoBSA}
\title{Check fitted RoBSA object for errors and warnings}
\usage{
check_RoBSA(fit)
}
\arguments{
\item{fit}{a fitted RoBSA object.}
}
\value{
\code{check_RoBSA} returns a vector of error and
warning messages.
}
\description{
Checks fitted RoBSA object
for warnings and errors and prints them to the
console.
}
