% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{exp-aft}
\alias{exp-aft}
\alias{exp_aft_log_density}
\alias{exp_aft_log_hazard}
\alias{exp_aft_log_survival}
\alias{exp_aft_density}
\alias{exp_aft_hazard}
\alias{exp_aft_survival}
\alias{exp_aft_mean}
\alias{exp_aft_sd}
\alias{exp_aft_r}
\alias{exp_aft_q}
\alias{exp_aft_p}
\title{Exponential AFT parametric family.}
\usage{
exp_aft_log_density(t, eta)

exp_aft_log_hazard(t, eta)

exp_aft_log_survival(t, eta)

exp_aft_density(t, eta)

exp_aft_hazard(t, eta)

exp_aft_survival(t, eta)

exp_aft_mean(eta)

exp_aft_sd(eta)

exp_aft_r(n, eta)

exp_aft_q(p, eta)

exp_aft_p(q, eta)
}
\arguments{
\item{t}{vector of survival times}

\item{eta}{linear predictor}

\item{n}{number of observations}

\item{p}{vector of probabilities}

\item{q}{vector of quantiles}

\item{location}{vector of locations.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities
\code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities
are \eqn{P[X \le x]}, otherwise, \eqn{P[X \ge x]}.}
}
\description{
(log) density, hazard, and survival
functions for AFT exponential parametric family.
}
