% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurv-tools.R
\name{extract_flexsurv}
\alias{extract_flexsurv}
\title{Extract parameter estimates from \code{flexsurv} object}
\usage{
extract_flexsurv(fit)
}
\arguments{
\item{fit}{an object fitted with the \code{flexsurv::flexsurvreg}
function}
}
\value{
\code{extract_flexsurv} return list of estimates lists for each
parameter.
}
\description{
\code{extract_flexsurv} extracts estimates from a
\code{flexsurv} object in and transform them to match the
\code{RoBSA} output.
}
