% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gamma-aft}
\alias{gamma-aft}
\alias{gamma_aft_log_density}
\alias{gamma_aft_log_hazard}
\alias{gamma_aft_log_survival}
\alias{gamma_aft_density}
\alias{gamma_aft_hazard}
\alias{gamma_aft_survival}
\alias{gamma_aft_mean}
\alias{gamma_aft_sd}
\alias{gamma_aft_r}
\alias{gamma_aft_q}
\alias{gamma_aft_p}
\title{Gamma AFT parametric family.}
\usage{
gamma_aft_log_density(t, eta, shape)

gamma_aft_log_hazard(t, eta, shape)

gamma_aft_log_survival(t, eta, shape)

gamma_aft_density(t, eta, shape)

gamma_aft_hazard(t, eta, shape)

gamma_aft_survival(t, eta, shape)

gamma_aft_mean(eta, shape)

gamma_aft_sd(eta, shape)

gamma_aft_r(n, eta, shape)

gamma_aft_q(p, eta, shape)

gamma_aft_p(q, eta, shape)
}
\arguments{
\item{t}{vector of survival times}

\item{eta}{linear predictor}

\item{n}{number of observations}

\item{p}{vector of probabilities}

\item{q}{vector of quantiles}

\item{sd}{auxiliary parameter}

\item{location}{vector of locations.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities
\code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities
are \eqn{P[X \le x]}, otherwise, \eqn{P[X \ge x]}.}
}
\description{
(log) density, hazard, and survival
functions for AFT gamma parametric family.
}
