% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{is.RoBSA}
\alias{is.RoBSA}
\title{Reports whether x is a RoBSA object}
\usage{
is.RoBSA(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
is.RoBSA returns a boolean.
}
\description{
Reports whether x is a RoBSA object
}
