% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lnorm-aft}
\alias{lnorm-aft}
\alias{lnorm_aft_log_density}
\alias{lnorm_aft_log_hazard}
\alias{lnorm_aft_log_survival}
\alias{lnorm_aft_density}
\alias{lnorm_aft_hazard}
\alias{lnorm_aft_survival}
\alias{lnorm_aft_mean}
\alias{lnorm_aft_sd}
\alias{lnorm_aft_r}
\alias{lnorm_aft_q}
\alias{lnorm_aft_p}
\title{Log-normal AFT parametric family.}
\usage{
lnorm_aft_log_density(t, eta, sd)

lnorm_aft_log_hazard(t, eta, sd)

lnorm_aft_log_survival(t, eta, sd)

lnorm_aft_density(t, eta, sd)

lnorm_aft_hazard(t, eta, sd)

lnorm_aft_survival(t, eta, sd)

lnorm_aft_mean(eta, sd)

lnorm_aft_sd(eta, sd)

lnorm_aft_r(n, eta, sd)

lnorm_aft_q(p, eta, sd)

lnorm_aft_p(q, eta, sd)
}
\arguments{
\item{t}{vector of survival times}

\item{eta}{linear predictor}

\item{sd}{auxiliary parameter}

\item{n}{number of observations}

\item{p}{vector of probabilities}

\item{q}{vector of quantiles}

\item{location}{vector of locations.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities
\code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities
are \eqn{P[X \le x]}, otherwise, \eqn{P[X \ge x]}.}
}
\description{
(log) density, hazard, and survival
functions for AFT log-normal parametric family.
}
