% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prediction}
\alias{plot_prediction}
\alias{plot_survival}
\alias{plot_hazard}
\alias{plot_density}
\title{Survival plots for a RoBSA object}
\usage{
plot_prediction(
  x,
  type = "survival",
  time_range = NULL,
  new_data = NULL,
  predictor = NULL,
  covariates_data = NULL,
  conditional = FALSE,
  plot_type = "base",
  samples = 10000,
  ...
)

plot_survival(
  x,
  time_range = NULL,
  new_data = NULL,
  predictor = NULL,
  covariates_data = NULL,
  conditional = FALSE,
  plot_type = "base",
  samples = 10000,
  ...
)

plot_hazard(
  x,
  time_range = NULL,
  new_data = NULL,
  predictor = NULL,
  covariates_data = NULL,
  conditional = FALSE,
  plot_type = "base",
  samples = 10000,
  ...
)

plot_density(
  x,
  time_range = NULL,
  new_data = NULL,
  predictor = NULL,
  covariates_data = NULL,
  conditional = FALSE,
  plot_type = "base",
  samples = 10000,
  ...
)
}
\arguments{
\item{x}{a fitted RoBSA object.}

\item{type}{what type of prediction should be created}

\item{new_data}{a data.frame containing fully specified predictors for which
predictions should be made}

\item{predictor}{an alternative input to \code{new_data} that automatically
generates predictions for each level of the predictor across all either across
levels of covariates specified by \code{covariates_data} or at the default values
of other predictors}

\item{covariates_data}{a supplementary input to \code{predictor} that specifies
levels of covariates for which predictions should be made}

\item{conditional}{whether only models assuming presence of the specified
\code{predictor} should be used}

\item{samples}{number of posterior samples to be evaluated}

\item{...}{additional arguments.}
}
\description{
Survival plots for a RoBSA object
}
