% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.RoBSA}
\alias{predict.RoBSA}
\title{Plots a fitted RoBSA object}
\usage{
\method{predict}{RoBSA}(
  object,
  newdata,
  type = c("survival", "hazard", "density", "mean", "sd"),
  summarize = TRUE,
  models = FALSE,
  conditional = FALSE,
  samples = 10000
)
}
\arguments{
\item{x}{a fitted RoBSA object.}

\item{...}{additional arguments.}
}
\description{
Plots a fitted RoBSA object
}
