% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.RoBSA.prior}
\alias{print.RoBSA.prior}
\title{Prints a RoBSA.prior object}
\usage{
\method{print}{RoBSA.prior}(x, ...)
}
\arguments{
\item{x}{a RoBSA prior}

\item{...}{additional arguments
\describe{
\item{silent}{to silently return the print message.}
\item{plot}{to return \link[base]{bquote} formatted
prior name for plotting.}
\item{digits_estimates}{number of decimals to be displayed
for printed parameters.}
}}
}
\description{
Prints a RoBSA.prior object
}
\seealso{
\code{\link[=prior]{prior()}}
}
