% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.RoBSA}
\alias{print.RoBSA}
\title{Prints a fitted RoBSA object}
\usage{
\method{print}{RoBSA}(x, ...)
}
\arguments{
\item{x}{a fitted RoBSA object.}

\item{...}{additional arguments.}
}
\description{
Prints a fitted RoBSA object
}
\seealso{
\code{\link[=RoBSA]{RoBSA()}}
}
