% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.RoBSA}
\alias{print.summary.RoBSA}
\title{Prints summary object for RoBSA method}
\usage{
\method{print.summary}{RoBSA}(x, ...)
}
\arguments{
\item{x}{a summary of a RoBSA object}

\item{...}{additional arguments}
}
\description{
Prints summary object for RoBSA method
}
\seealso{
\code{\link[=RoBSA]{RoBSA()}}
}
