% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoBSA-package.R
\docType{package}
\name{RoBSA-package}
\alias{RoBSA-package}
\alias{_PACKAGE}
\alias{RoBSA_package}
\alias{RoBSA.package}
\title{RoBSA: Robust Bayesian survival analysis}
\description{
Bayesian model-averaged parametric survival analysis
with ability to specify informed prior distributions and draw inference with
inclusion Bayes factors. See \insertCite{bartos2021informed;textual}{RoBSA}
for more details about the methodology.
}
\section{User guide}{
 See \insertCite{bartos2021informed;textual}{RoBSA},
for details regarding the RoBSA methodology.
}

\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbartos.github.io/RoBSA/}
  \item Report bugs at \url{https://github.com/FBartos/RoBSA/issues}
}

}
\author{
František Bartoš \email{f.bartos96@gmail.com}
}
\keyword{package}
