% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RoBSA}
\alias{RoBSA}
\title{Fit Robust Bayesian Survival Analysis}
\usage{
RoBSA(
  formula,
  data,
  priors = NULL,
  test_predictors = NULL,
  distributions = c("exp-aft", "weibull-aft", "lnorm-aft", "llogis-aft", "gamma-aft"),
  distributions_odds = rep(1, length(distributions)),
  prior_beta_null = get_default_prior_beta_null(),
  prior_beta_alt = get_default_prior_beta_alt(),
  prior_intercept = get_default_prior_intercept(),
  prior_aux = get_default_prior_aux(),
  chains = 2,
  iter = 5000,
  burnin = 1000,
  thin = 1,
  parallel = FALSE,
  control = NULL,
  save = "all",
  seed = NULL
)
}
\arguments{
\item{formula}{formula for the survival model}

\item{data}{data frame containing the data}

\item{test_predictors}{vector of predictor names
to be tested with Bayesian model-averaged testing.
Defaults to \code{NULL}, no parameters are tested.}

\item{distributions}{distributions of parametric
survival models}

\item{distributions_odds}{prior odds for the competing
distributions}

\item{prior_beta_null}{named list containing null prior
distribution for the predictors (with names corresponding
to the predictors)}

\item{prior_beta_alt}{named list containing prior
distribution for the predictors (with names corresponding
to the predictors)}

\item{prior_intercept}{named list containing prior
distribution for the intercepts (with names corresponding
to the distributions)}

\item{prior_aux}{named list containing prior
distribution for the auxiliary parameters (with names corresponding
to the distributions)}

\item{chains}{number of chains}

\item{iter}{number of sampling iterations}

\item{burnin}{number of burnin iterations}

\item{thin}{thinning of the chains}

\item{parallel}{whether the models should be fit in parallel}

\item{...}{additional arguments.}
}
\description{
Fits a RoBSA model. Please note
that the fitting function only supports dummy
coded predictors and cannot deal with factors.
}
