% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RoBSA_options}
\alias{RoBSA_options}
\alias{RoBSA.options}
\alias{RoBSA.get_option}
\title{Options for the RoBSA package}
\usage{
RoBSA.options(...)

RoBSA.get_option(name)
}
\arguments{
\item{...}{named option(s) to change - for a list of available options, see
details below.}

\item{name}{the name of the option to get the current value of - for a list of
available options, see details below.}
}
\value{
The current value of all available RoBSA options (after applying any
changes specified) is returned invisibly as a named list.
}
\description{
A placeholder object and functions for the RoBSA package.
(adapted from the runjags R package).
}
