% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_calibration.R
\name{calibrate_meta_analytic}
\alias{calibrate_meta_analytic}
\title{Create meta-analytic predictive prior distributions}
\usage{
calibrate_meta_analytic(
  datasets,
  distributions = c("exp-aft", "weibull-aft", "lnorm-aft", "llogis-aft", "gamma-aft"),
  mu_scale = 100,
  tau_scale = 10
)
}
\arguments{
\item{datasets}{list of data.frames containing the historical
data. Each data.frame must contain a column named \code{"time"}
with the survival times and a column named \code{"status"}
with the censoring status.}

\item{distributions}{vector of parametric families for which
prior distributions ought to be calibrated}

\item{mu_scale}{scale of the prior Cauchy distribution for the meta-analytic
mean parameter}

\item{tau_scale}{scale of the prior Cauchy distribution for the meta-analytic
heterogeneity parameter}
}
\description{
Calibrates prior distributions for parametric
survival analysis based on historical data.
Returns a list of prior distribution for the intercepts
and auxiliary parameters.
}
