% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{contr.orthonormal}
\alias{contr.orthonormal}
\title{Orthornomal contrast matrix}
\usage{
contr.orthonormal(n, contrasts = TRUE)
}
\arguments{
\item{n}{a vector of levels for a factor, or the number of levels}

\item{contrasts}{logical indicating whether contrasts should be computed}
}
\value{
A matrix with n rows and k columns, with k = n - 1 if \code{contrasts = TRUE} and k = n
if \code{contrasts = FALSE}.
}
\description{
Return a matrix of orthornomal contrasts.
Code is based on \code{stanova::contr.bayes} and corresponding to description
by \insertCite{rouder2012default;textual}{BayesTools}
}
\examples{
contr.orthonormal(c(1, 2))
contr.orthonormal(c(1, 2, 3))

}
\references{
\insertAllCited{}
}
