% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gamma-aft}
\alias{gamma-aft}
\alias{gamma_aft_log_density}
\alias{gamma_aft_log_hazard}
\alias{gamma_aft_log_survival}
\alias{gamma_aft_density}
\alias{gamma_aft_hazard}
\alias{gamma_aft_survival}
\alias{gamma_aft_mean}
\alias{gamma_aft_sd}
\alias{gamma_aft_r}
\alias{gamma_aft_q}
\alias{gamma_aft_p}
\title{Gamma AFT parametric family.}
\usage{
gamma_aft_log_density(t, eta, shape)

gamma_aft_log_hazard(t, eta, shape)

gamma_aft_log_survival(t, eta, shape)

gamma_aft_density(t, eta, shape)

gamma_aft_hazard(t, eta, shape)

gamma_aft_survival(t, eta, shape)

gamma_aft_mean(eta, shape)

gamma_aft_sd(eta, shape)

gamma_aft_r(n, eta, shape)

gamma_aft_q(p, eta, shape)

gamma_aft_p(q, eta, shape)
}
\arguments{
\item{t}{vector of survival times}

\item{eta}{linear predictor}

\item{shape}{auxiliary parameter}

\item{n}{number of observations}

\item{p}{vector of probabilities}

\item{q}{vector of quantiles}
}
\value{
\code{gamma_aft_density}, \code{gamma_aft_hazard}, and
\code{gamma_aft_survival} return the density, hazard, and survival
of the specified survival distribution. The \code{gamma_aft_log_density},
\code{gamma_aft_log_hazard}, \code{gamma_aft_log_survival} return log of
the corresponding qualities. \code{gamma_aft_mean} and \code{gamma_aft_sd}
return the mean and standard deviation of the specified survival distribution.
\code{gamma_aft_r}, \code{gamma_aft_q}, and \code{gamma_aft_p} return a random
generation, quantiles, and cumulative probabilities of the specified
survival distribution.
}
\description{
(log) density, hazard, and survival
functions for AFT gamma parametric family.
}
