% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.RoBSA}
\alias{plot.RoBSA}
\title{Plots a fitted RoBSA object}
\usage{
\method{plot}{RoBSA}(x, ...)
}
\arguments{
\item{x}{a fitted RoBSA object.}

\item{...}{additional arguments.}
}
\description{
Plots a fitted RoBSA object
}
