% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_survival}
\alias{plot_survival}
\title{Plots a fitted RoBSA object}
\usage{
plot_survival(x, ...)
}
\arguments{
\item{x}{a fitted RoBSA object.}

\item{...}{additional arguments.}
}
\description{
Plots a fitted RoBSA object
}
