% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoBSA-package.R
\docType{package}
\name{RoBSA-package}
\alias{RoBSA-package}
\alias{_PACKAGE}
\title{RoBSA: Robust Bayesian survival analysis}
\description{
A framework for estimating ensembles of parametric survival models
and using Bayesian model averaging to combine them. The
ensembles use Bayes factors to test for the presence or absence of the
individual components (e.g., treatment effect, parametric family, etc...) and
model-averages parameter estimates based on posterior model probabilities.
The user can define a wide range of informative priors for all parameters
of interest. The package provides convenient functions for summary, visualizations,
fit diagnostics, and prior distribution calibration.
}
\details{
\code{\link{RoBSA}} fits parametric models ....
}
\section{User guide}{
 The \bold{RoBSA user guide} vignette explains the
methods in detail, and gives several worked examples.  A further vignette
\bold{RoBSA-examples} gives a few more complicated examples, and users
are encouraged to submit their own.
}

\references{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbartos.github.io/RoBSA/}
  \item Report bugs at \url{https://github.com/FBartos/RoBSA/issues}
}

}
\author{
František Bartoš \email{f.bartos96@gmail.com}
}
\keyword{package}
